refactor(test): Factor out event functions for HubPoolClient
This slims down the MockSpokePoolClient implementation and makes the
generic implementation available for use in the MockHubPoolClient.

Also:
 - Support generating "EnabledDepositRoute" events.
 - Cleanup incorrect isSlowRelay member from old fill events.
 - Fill out placeholder SpokePool event signatures and topics.
 - Export MockSpokePoolClient from test/utils.
