+++ test/mocks/index.ts
@@ -2,6 +2,7 @@ export * from "./MockBundleDataClient";
 export * from "./MockHubPoolClient";
 export * from "./MockProfitClient";
 export * from "./MockAdapterManager";
+export * from "./MockSpokePoolClient";
 export * from "./MockTokenClient";
 export * from "./MockTransactionClient";
 export * from "./MockInventoryClient";
